Finalized Yolo
There are 2 versions, yolo.py (no visualization, faster) and yolo_vis.py (visualization, both depth and color).
The color should be 100% accurate and depth is accurate to the centimeter. 
Both include a function to choose which types of data to return, based on mogo color, all mogos or closest mogo or any combination.